CCSD-5009: Aggregate Signbank fields to prevent Dictionary duplicates (#192)
## JIRA Ticket

[CCSD-5009 NZSL: videos not displaying correctly in NZSL Online
Dictionary](https://ackama.atlassian.net/browse/CCSD-5009)

## Changes

- Postgres aggregation added to database columns `related_to` and
`usage`
- Superfluous comments removed


## Notes

This started in NZSL Dictionary, because duplicate example videos were
appearing for certain signs.
Eg. `attitude`, `expensive`, `visit` and a number of others.
At first it was suspected it was the fault of the script that pulls data
from NZSL Signbank.
This creates `CSV` files, and from them creates the `nzsl.db` `sqlite`
file, that ends up in `AWS S3` , and that the Dictionary is built from.

It transpired the fault was actually with Signbank, and the way
`GlossListView` `subquery_render_to_csv_response()` was aggregating the
postgres data before returning it.

Two columns were not being aggregated:

-  `related_to` - this is mediated via the `RelationToForeignSign` model
- `usage` - a `FieldChoice` many-to-many column

Further notes:
- `related_to` returns the `RelationToForeignSign` model value field
`other_lang`, but there is another field `other_lang_gloss` in the model
that in a perfect world would be added also. The client said it was Ok
to ignore this second field, but one day somebody might figure out how
to include it.
- No other aggregation misses have been spotted, so far. Essentially
there should never be multiple rows in the output CSV with the same
Signbank ID number.