CC-2138 NZSL Upgrade Django 3.2.25 > 4.2 (#170)
## JIRA Ticket

[CC-2138 NZSL: upgrade signbank to django
4.2](https://ackama.atlassian.net/browse/CC-2138)

## Changes

- Django 3.2.25 -> ~4.2.13
 
Upstream changes:

- `ugettext, ugettext_lazy -> gettext, gettext_lazy`
- `dictionary/paginate.html`, conditional syntax changed: `ifnoteqal ->
if !`
- Workaround for an issue in `django-tagging` which seems to no longer
be maintained, by aliasing `smart_text -> smart_str`
- `signbank/video/views.py`, function no longer available `is_ajax()`,
replaced by header conditional

Pinnings so that `collectstatic` runs -
Several modules depended on an old version of the `Signal` class,
brought in via `notifications`.
This results in a break thus
``` TypeError: __init__() got an unexpected keyword argument 'providing_args' ```
An example of it in action is [here](https://github.com/django-notifications/django-notifications/issues/322)
Allowing these modules to upgrade brings in only the later, compatible version of `notifications`, and `collectstatic` completes without error.
I pinned these to the lowest version from a 'change as little as possible' standpoint, but from a security standpoint we possibly want to drive these as high as compatibility will allow. The standing suggestion is that, post this upgrade, we break them into their own individual upgrade PRs to address this.

- Pinning `django-reversion` to lowest compatible version: `4.0.0 -> 4.0.2`
- Pinning `django-contrib-comments` to lowest compatible version: `1.9.2 -> 2.0.0`
- Pinning `django-notifications-hq` to lowest compatible version: `1.6.0 -> 1.7.0`
- Pinning `django-summernote` to lowest compatible version: `0.8.11.6 -> 0.8.20.0`

Misc:

- Upgraded `django-contrib-comments` to latest `2.2.0` to fix missing migration bug
- Removed deprecated `version` string from `docker-compose.yml` to fix warning