CC-2305: create management commands for healing Video file S3<->Postgres relationships (#180)
Adds management commands to resolve discrepancies between NZSL
Signbank's Postgres database and Amazon S3 file storage.

## Jira ticket
[CC-2305 NZSL Signbank: UAT database
backport](https://ackama.atlassian.net/browse/CC-2305)

## Changes

- Three management commands created, details below.
- README documentation created.

These management commands are used to report on relationships between
Signbank's Postgres database and Amazon's S3 file storage, and then
_assist_ with effecting some types of repair where discrepancies exist.
See documentation.

These commands only perform certain actions, other operations have to be
manually commanded using AWS `cli` or other means, using the output from
these commands as data. See documentation.

- The commands use the `Boto3` python library to talk to AWS S3.
- They use an external client to talk to Postgres.
- They output diagnostic and progress information on STDERR. All data
output is on STDOUT and may be safely redirected.

The commands require, usually in the environment:

- An AWS profile - eg. `AWS_PROFILE` environment variable set to a
pre-configured profile.
- A Postgres context - eg. `DATABASE_URL` environment variable with
target and credentials.

The commands have some common arguments:

- `--help` or `-h` - emit a Help message showing the available
arguments.
- `--env` - specifies the target environment, eg. `dev`, `uat`,
`production`. This is used to contruct the name of the AWS S3 bucket
name, eg. `nzsl-signbank-media-uat`. The default is `uat`.
- `--pgcli` - allows the user to specify a different path for the
Postgres command-line client. The default is `/usr/bin/psql`.

<br />

### Commands

<br />

- `get-video-s3-acls`

This command produces a full report on NZSL vs S3.
It outputs as CSV, with headers.

<br />

- `find-fixable-s3-orphans`

This command accesses the database and S3 in a similar way to
`get-video-s3-acls.py`.

It finds S3 objects that have no corresponding NZSL Signbank database
record. These are 'orphaned' S3 objects.
It then parses the name string of the object and attempts to find an
NZSL Signbank record that matches it. This is not guaranteed to be
correct, so the output needs human review.
It outputs what it finds as CSV with header, in a format that can be
digested by the 3rd command `repair-fixable-s3-orphans.py`.

<br />

- `repair-fixable-s3-orphans`

This attempts to unify NZSL Signbank records with S3 orphans, by
digesting a CSV input of the same format as output by
`find-fixable-orphans.py`. It does this by generating `GlossVideo`
Django objects where necessary, and associating them with the correct
`Gloss` Django objects. This operation _changes_ the database contents
and so must be used with caution.

<br />